+++ R/helpers.R
@@ -1,0 +1,1 @@
+done <- function(...) cli::cat_bullet(..., bullet = "tick", bullet_col = "green")